Fix some warnings

+++ game/src/gameState/options/VideoOptionsState.c
@@ -7,7 +7,6 @@
 #include <engine/graphics/Font.h>
 #include <engine/graphics/RenderingHelpers.h>
 #include <engine/helpers/BackgroundMapManager.h>
-#include <engine/physics/MapPhysics.h>
 #include <engine/structs/Color.h>
 #include <engine/structs/GameState.h>
 #include <engine/structs/GlobalState.h>
@@ -145,7 +144,7 @@ void SldOptionsLod(const float value)
 
 void SldOptionsMaxFps(const float value)
 {
-	GetState()->options.maxFps = value;
+	GetState()->options.maxFps = (uint16_t)value;
 }
 
 void SldOptionsFov(const float value)
